+++ AspNetCore/AT.Common.AspNetCore.Publish/CHANGELOG.md
@@ -19,6 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security <!-- in case of vulnerabilities. -->
 
+## 1.2.1
+
+### Added
+
+- feat(extensions): added enum to string converter as default behavior when adding controllers
+- feat(extensions): added custom controller detector to also detect internal controllers (before it detected only public classes)
+
 ## 1.2.0
 
 ### Added